Title(feat)：升级PyQt5至5.15.11版本。 Cause：版本编译时选择的最新版本。未后续功能扩展和bug修改做准备。 Solution：添加依赖说明文件。修改resize等参数类型发生变化的API传参。 Effect: N/A Reviewer: N/A 额外说明：N/A

+++ XulDebugTool/ui/widget/DataQueryDialog.py
@@ -90,8 +90,8 @@ class DataQueryDialog(BaseDialog):
                     row += 1
 
         self.tableView = QtWidgets.QTableWidget(row,2,self)
-        self.tableView.move(36, 3*Utils.getItemHeight())
-        self.tableView.resize(300, Utils.getItemHeight() * (row+1))
+        self.tableView.move(36, int(3*Utils.getItemHeight()))
+        self.tableView.resize(300, int(Utils.getItemHeight() * (row+1)))
         self.tableView.setColumnWidth(0, 150)
         self.tableView.setColumnWidth(1, 150)
         self.tableView.horizontalHeader().setFixedHeight(Utils.getItemHeight())
@@ -149,9 +149,9 @@ class DataQueryDialog(BaseDialog):
                 self.tableView.setItem(row, 1, QtWidgets.QTableWidgetItem())
                 # 加上表头,不超过7行,超过滚动条显示
                 if (self.currentRowCount + 2) <= 7:
-                    self.tableView.resize(300, Utils.getItemHeight() * (self.currentRowCount + 2))
+                    self.tableView.resize(300, int(Utils.getItemHeight() * (self.currentRowCount + 2)))
                 else:
-                    self.tableView.resize(322, Utils.getItemHeight() * 7)
+                    self.tableView.resize(322, int(Utils.getItemHeight() * 7))
         url = self.url + self.__getQueryParam()
         self.requestLineEdit.setText(url)
